+++ tests/test_07_render/test_708b_path_tools.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2020, Manfred Moitzi
+#  Copyright (c) 2020-2022, Manfred Moitzi
 #  License: MIT License
 import pytest
 import math
@@ -21,6 +21,8 @@ from ezdxf.path import (
     from_vertices,
     to_multi_path,
     single_paths,
+    lines_to_curve3,
+    lines_to_curve4,
 )
 from ezdxf.path import make_path, Command
 from ezdxf.entities import BoundaryPathType, EdgeType
@@ -612,3 +614,113 @@ def test_issue_494_make_path_from_spline_defined_by_fit_points_and_tangents():
     ]
     p = make_path(spline)
     assert len(p) > 0
+
+
+class TestAllLinesToCurveConverter:
+    def test_create_a_curve3_command(self):
+        path = Path()
+        path.line_to((1, 0))
+        path = lines_to_curve3(path)
+        assert path[0].type == Command.CURVE3_TO
+
+    def test_create_a_curve4_command(self):
+        path = Path()
+        path.line_to((1, 0))
+        path = lines_to_curve4(path)
+        assert path[0].type == Command.CURVE4_TO
+
+    @pytest.mark.parametrize(
+        "func",
+        [
+            lines_to_curve3,
+            lines_to_curve4,
+        ],
+    )
+    def test_line_to_curve_creates_a_linear_segment(self, func):
+        v1, v2 = 1, 2
+        path = Path(start=(v1, v1, v1))
+        path.line_to((v2, v2, v2))
+        path = func(path)
+        vertices = list(path.flattening(1))
+        assert len(vertices) > 2
+        assert all(
+            [
+                math.isclose(v.x, v.y) and math.isclose(v.x, v.z)
+                for v in vertices
+            ]
+        ), "all vertices have to be located along a line (x == y == z)"
+
+    def test_remove_line_segments_of_zero_length_at_the_start(self):
+        # CURVE3_TO and CURVE4_TO can not process zero length segments
+        path = Path()
+        path.line_to((0, 0))  # line segment of length==0 should be removed
+        path.line_to((1, 0))
+        path = lines_to_curve4(path)
+        assert len(path) == 1
+        assert path.start == (0, 0)
+        assert path[0].type == Command.CURVE4_TO
+        assert path[0].end == (1, 0)
+
+    def test_remove_line_segments_of_zero_length_between_commands(self):
+        # CURVE3_TO and CURVE4_TO can not process zero length segments
+        path = Path()
+        path.line_to((1, 0))
+        path.line_to((1, 0))  # line segment of length==0 should be removed
+        path.line_to((2, 0))
+        path = lines_to_curve4(path)
+        assert len(path) == 2
+        assert path.start == (0, 0)
+        assert path[0].type == Command.CURVE4_TO
+        assert path[0].end == (1, 0)
+        assert path[1].type == Command.CURVE4_TO
+        assert path[1].end == (2, 0)
+
+    def test_remove_line_segments_of_zero_length_at_the_end(self):
+        # CURVE3_TO and CURVE4_TO can not process zero length segments
+        path = Path()
+        path.line_to((1, 0))
+        path.line_to((1, 0))  # line segment of length==0 should be removed
+        path = lines_to_curve4(path)
+        assert len(path) == 1
+        assert path.start == (0, 0)
+        assert path[0].type == Command.CURVE4_TO
+        assert path[0].end == (1, 0)
+
+    def test_does_not_remove_a_line_representing_a_single_point(self):
+        path = Path((1, 0))
+        path.line_to((1, 0))  # represents the point (1, 0)
+        path = lines_to_curve4(path)
+        assert len(path) == 1
+        assert path[0].type == Command.LINE_TO
+
+    @pytest.mark.parametrize(
+        "start,delta",
+        [
+            (0, 1e-11),  # uses absolute tolerance of 1e-12 near zero!
+            (10, 1e-8),  # uses relative tolerance of 1e-9 away from zero!
+        ],
+    )
+    def test_for_very_short_line_segments(self, start, delta):
+        path = Path((start, 0, 0))
+        path.line_to((start + delta, 0, 0))
+        path = lines_to_curve4(path)
+        assert len(path) == 1
+        assert path[0].type == Command.CURVE4_TO
+        assert len(list(path.flattening(1))) > 3
+
+    @pytest.mark.parametrize(
+        "start,delta",
+        [
+            (0, 1e-12),  # uses absolute tolerance of 1e-12 near zero!
+            (10, 1e-9),  # uses relative tolerance of 1e-9 away from zero!
+        ],
+    )
+    def test_which_length_is_too_short_to_create_a_curve(self, start, delta):
+        path = Path((start, 0, 0))
+        path.line_to((start + delta, 0, 0))
+        path = lines_to_curve4(path)
+        assert len(path) == 1
+        assert (
+            path[0].type == Command.LINE_TO
+        ), "should not remove a single line segment representing a point"
+        assert len(list(path.flattening(1))) == 2