+++ NEWS.md
@@ -45,7 +45,10 @@ Version 0.18b1 - dev
 - CHANGE: `recover` module - recovered integer and float values are logged as severe errors
 - CHANGE: method `Path.all_lines_to_curve3` replaced by function `path.lines_to_curve3()`
 - CHANGE: method `Path.all_lines_to_curve4` replaced by function `path.lines_to_curve4()`
-- BUGFIX: add missing caret decoding to `fast_plain_mtext()` [#620](https://github.com/mozman/ezdxf/issues/620)
+- BUGFIX [#640](https://github.com/mozman/ezdxf/issues/640): 
+  DXF loader ignore data beyond `EOF` tag 
+- BUGFIX [#620](https://github.com/mozman/ezdxf/issues/620): 
+  add missing caret decoding to `fast_plain_mtext()` 
 
 Version 0.17.2 - 2022-01-06
 ---------------------------